Move redis dependency into worker
