+++ docker/worker/Pipfile
@@ -11,6 +11,7 @@ google-cloud-ndb = "*"
 pygit2 = "==1.5.0"
 PyYAML = "*"
 google-cloud-storage = "*"
+redis = "*"
 
 [dev-packages]
 yapf = "*"
